S07 Enroll to/exit course - Complete
Completed S07 as far as possible. Users can enroll to course, which
causes the user to become course participant. Technically, part of S07
was to send notification to teacher about user wanting to enroll, but I
believe it's better to postpone this to S17 Add/accept student to
course. Students can freely exit a course. Teachers can exit provided
they are not the only teacher in the course.

Also added enhancements to db.php for transaction support. Added
transactions when creating a course to prevent issues with course being
created without teacher.